moved props to types.ts
