+++ packages/suite-base/src/panels/StateTransitions/types.ts
@@ -5,6 +5,7 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
+import { SaveConfig } from "@lichtblick/suite-base/types/panels";
 import { TimestampMethod } from "@lichtblick/suite-base/util/time";
 
 export type StateTransitionPath = {
@@ -22,4 +23,11 @@ export type StateTransitionConfig = {
   xAxisMinValue?: number;
   xAxisRange?: number;
   showPoints?: boolean;
+};
+
+export type PathLegendProps = {
+  paths: StateTransitionPath[];
+  heightPerTopic: number;
+  setFocusedPath: (value: string[] | undefined) => void;
+  saveConfig: SaveConfig<StateTransitionConfig>;
 };
